Allow keeping Cygwin out of PATH
In some circumstances, it might be useful to have Cygwin installed, but
not have its executables take precedence over pre-installed packages.
Add an "add-to-path" option that, if set to "false", will skip the step
of adding Cygwin's /bin directory to PATH.

Also update documentation to mention this option, and update the test
scripts to check it works as expected; the default behaviour is tested
as part of the "install-dir-test", and the new behaviour as part of the
main test.