Allow save-only/restore-only cache runs
Add an option to restore a cache without creating a new one, or to
create a cache without restoring an old one.  This is useful in the
scenario where the action is called multiple times in a run, as it
allows the cache to only be restored on the first call, and for a new
cache to only be written on the last call.

This option replaces the `package-cache-key` option; as discussed in #6
that option isn't very useful.

Additionally, add some more scopes to the cache name.  Without this
change, a second attempt to create a cache in the same action run (e.g.
because one step installs some packages, then another step installs some
more) will fail because of the cache name collision.

Also update the README and tests, although the tests are only getting a
very quick update for this function, as I'm going to do a much more
comprehensive test update separately.