Move the install script to a dedicated file
This allows developers to work with syntax highlighting;
it allows linters to check for syntax errors and style issues;
and it allows for the development of a test suite
using Powershell tools like e.g. Pester.

As a part of this change, `${{ inputs.* }}` references were replaced.
The script now relies on environment variables named `inputs_*`.